Fix tel: URI scheme handler
This adds a new IntentHelper method to process intents, extract the tel: URI scheme phone number if one is found in the intent, and return the appropriate (Empty or TelUriScheme) return value.
This also introduces RobolectricRunner because Uri is an Android-specific class and its creation in unit tests led to "not mocked" errors.