Update "Share dialog" intent filter
This updates the existing intent filter to register to handle text/plain shares.
This allows the app to receive shares and process the shared content to extract a phone number.

At the moment, this assumes the happy path that the shared content will be a tel:<phone number with country code>. A future commit will add error handling and parsing of the given text to regex-extract phone numbers and tel: links amidst other content.