Refactor intent processing into a use case
As additional scenarios for intent processing are encountered (which I expect, such as the sms URI scheme whose handling was requested), the intent processing code could become unwieldy to read/test. In preparation, this refactors intent processing into a separate use case that is thoroughly tested.

I believe this also addresses a bug in intent handling related to sharing copied text with the app.