Add VCard/"share contact" support
This adds an intent filter to register as a potential option to open shared VCF files from third-party apps.
The shared VCF file is examined for phone numbers, and all phone numbers are inserted into the phone number input field.

Out of scope for this PR: When multiple phone numbers are available, the "Open" buttons do not function. A future PR will show a dialog to prompt the user to select a phone number, since this scenario also applies to text shared with the app.