Update Instant creation to use millisecond precision
This adds a new Instant extension function that sets the precision of DateUtils-generated Instants to use millisecond precision (and not microsecond precision, as is the default).
The Instant extension addresses the problem when a test creates an Instant (with microsecond precision) which is then persisted into a SQLite database (which uses millisecond precision) and then compared against the original record, the timestamp would be flagged as not matching due to the precision mismatch.